server/ai/requirements.txt
torch
torchvision
Pillow
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git